+++ frontend/app/src/components/home/ChatEventList.svelte
@@ -36,10 +36,17 @@
     import { _ } from "svelte-i18n";
     import { pop } from "../../utils/transition";
     import { iconSize } from "../../stores/iconSize";
-    import { eventListLastScrolled, eventListScrollTop } from "../../stores/scrollPos";
+    import {
+        eventListLastScrolled,
+        eventListScrollTop,
+        eventListScrolling,
+        reverseScroll,
+    } from "../../stores/scrollPos";
+    import TimelineDate from "./TimelineDate.svelte";
 
     const FROM_BOTTOM_THRESHOLD = 600;
     const LOADING_THRESHOLD = 400;
+    const SCROLL_THRESHOLD = 500;
     const client = getContext<OpenChat>("client");
 
     export let rootSelector: string;
@@ -67,6 +74,7 @@
     let scrollingToMessage = false;
     let scrollToBottomOnSend = false;
     let destroyed = false;
+    let labelObserver: IntersectionObserver;
 
     $: failedMessagesStore = client.failedMessagesStore;
     $: threadSummary = threadRootEvent?.event.thread;
@@ -76,16 +84,42 @@
         clientHeight: messagesDiv!.clientHeight,
         scrollTop: messagesDiv!.scrollTop,
         fromBottom: fromBottom(),
+        fromTop: fromTop(),
         insideBottomThreshold: insideBottomThreshold(),
         insideTopThreshold: insideTopThreshold(),
         loadingFromScroll: loadingFromUserScroll,
         eventCount: events.length,
     });
 
+    const bottom = () => {
+        if (messagesDiv) {
+            if (reverseScroll) {
+                return 0;
+            } else {
+                return messagesDiv.scrollHeight - messagesDiv.clientHeight;
+            }
+        }
+        return 0;
+    };
+
     const fromBottom = () => {
         if (messagesDiv) {
-            const bottom = messagesDiv.scrollHeight - messagesDiv.clientHeight;
-            return bottom - messagesDiv.scrollTop;
+            if (reverseScroll) {
+                return -messagesDiv.scrollTop;
+            } else {
+                return bottom() - messagesDiv.scrollTop;
+            }
+        }
+        return 0;
+    };
+
+    const fromTop = () => {
+        if (messagesDiv) {
+            if (reverseScroll) {
+                return messagesDiv.scrollHeight - messagesDiv.clientHeight - fromBottom();
+            } else {
+                return messagesDiv.scrollTop;
+            }
         }
         return 0;
     };
@@ -95,10 +129,11 @@
     };
 
     const insideTopThreshold = () => {
-        return (messagesDiv?.scrollTop ?? 0) < LOADING_THRESHOLD;
+        return fromTop() < LOADING_THRESHOLD;
     };
 
     let showGoToBottom = false;
+    let floatingTimestamp: bigint | undefined = undefined;
 
     beforeUpdate(() => {
         previousScrollHeight = messagesDiv?.scrollHeight;
@@ -109,7 +144,39 @@
         showGoToBottom = fromBottom() > FROM_BOTTOM_THRESHOLD;
     });
 
+    function elementIsOffTheTop(el: Element): boolean {
+        return el.getBoundingClientRect().top < 95;
+    }
+
     onMount(() => {
+        const options = {
+            root: messagesDiv as Element,
+            rootMargin: "-15px 0px 0px 0px",
+            threshold: [0, 0.5, 1],
+        };
+
+        labelObserver = new IntersectionObserver((_entries: IntersectionObserverEntry[]) => {
+            const labels = [
+                ...document.querySelectorAll(".date-label[data-timestamp]:not(.floating)"),
+            ];
+            if (!reverseScroll) {
+                labels.reverse();
+            }
+            floatingTimestamp = undefined;
+            for (const label of labels) {
+                const float = elementIsOffTheTop(label);
+                if (float && floatingTimestamp === undefined) {
+                    (label as HTMLElement).style.opacity = "0";
+                    const timestamp = label.getAttribute("data-timestamp");
+                    if (timestamp != null) {
+                        floatingTimestamp = BigInt(timestamp);
+                    }
+                } else {
+                    (label as HTMLElement).style.opacity = "1";
+                }
+            }
+        }, options);
+
         if (messagesDiv !== undefined && $eventListScrollTop !== undefined && maintainScroll) {
             interruptScroll(() => {
                 if (messagesDiv !== undefined && $eventListScrollTop !== undefined) {
@@ -232,7 +299,7 @@
     ): Promise<void> {
         if (messagesDiv) {
             messagesDiv?.scrollTo({
-                top: messagesDiv.scrollHeight - messagesDiv.clientHeight,
+                top: bottom(),
                 behavior,
             });
         }
@@ -397,6 +464,39 @@
     async function onLoadedPreviousMessages(initialLoad: boolean) {
         await tick();
         await resetScroll(initialLoad);
+        if (reverseScroll) {
+            // Seems like we *must* interrupt the scroll to stop runaway loading
+            // even though we do not need to do any adjustment of the scrollTop in this direction.
+            // This seems to help on chrome but not on safari (God help us).
+            await interruptScroll(() => {
+                console.debug("SCROLL: onLoadedPrevious interrupt");
+            });
+        } else {
+            await adjustScrollTopIfNecessary(initialLoad, true);
+        }
+        await loadMoreIfRequired(loadingFromUserScroll, initialLoad);
+    }
+
+    async function onLoadedNewMessages() {
+        await tick();
+
+        if (reverseScroll) {
+            await adjustScrollTopIfNecessary(false, false);
+        }
+
+        if (
+            !loadingFromUserScroll &&
+            !client.moreNewMessagesAvailable(chat.id, threadRootEvent) &&
+            insideBottomThreshold()
+        ) {
+            // only scroll if we are now within threshold from the bottom
+            scrollBottom("smooth");
+        }
+
+        await loadMoreIfRequired(loadingFromUserScroll);
+    }
+
+    async function adjustScrollTopIfNecessary(initialLoad: boolean, add: boolean): Promise<void> {
         if (
             !initialLoad &&
             messagesDiv !== undefined &&
@@ -409,34 +509,22 @@
             const diffDiff = scrollHeightDiff - scrollTopDiff;
             // sometimes chrome is *a little* out but it we only want to intervene if if's way off
             if (diffDiff > sensitivityThreshold) {
-                interruptScroll(() => {
+                await interruptScroll(() => {
                     if (messagesDiv !== undefined && previousScrollTop !== undefined) {
-                        let adjusted = messagesDiv.scrollTop + scrollHeightDiff;
-                        // This is still not great on iphone particularly in the groups that have a high proportion of non-message events
+                        let adjusted = add
+                            ? messagesDiv.scrollTop + scrollHeightDiff
+                            : messagesDiv.scrollTop - scrollHeightDiff;
                         messagesDiv.scrollTop = adjusted;
                         console.debug("SCROLL: adjusted: ", {
                             ...keyMeasurements(),
                             scrollHeightDiff,
                             scrollTopDiff,
+                            reverseRender: reverseScroll,
                         });
                     }
                 });
             }
         }
-        await loadMoreIfRequired(loadingFromUserScroll, initialLoad);
-    }
-
-    async function onLoadedNewMessages() {
-        if (
-            !loadingFromUserScroll &&
-            !client.moreNewMessagesAvailable(chat.id, threadRootEvent) &&
-            insideBottomThreshold()
-        ) {
-            // only scroll if we are now within threshold from the bottom
-            await scrollBottom("smooth");
-        }
-
-        await loadMoreIfRequired(loadingFromUserScroll);
     }
 
     // this *looks* crazy - but the idea is that before we programmatically scroll the messages div
@@ -455,6 +543,7 @@
     }
 
     function onUserScroll() {
+        trackScrollStop(SCROLL_THRESHOLD);
         if (maintainScroll) {
             $eventListScrollTop = messagesDiv?.scrollTop;
         }
@@ -479,15 +568,28 @@
             loadIndexThenScrollToBottom(chat.latestMessage?.event.messageIndex ?? -1);
         }
     }
+
+    let scrollTimeout: number | undefined = undefined;
+    function trackScrollStop(delay: number) {
+        eventListScrolling.set(true);
+        clearTimeout(scrollTimeout);
+        scrollTimeout = window.setTimeout(() => {
+            eventListScrolling.set(false);
+        }, delay);
+    }
 </script>
 
+{#if floatingTimestamp !== undefined}
+    <TimelineDate observer={labelObserver} timestamp={BigInt(floatingTimestamp)} floating />
+{/if}
 <div
     bind:this={messagesDiv}
     bind:clientHeight={messagesDivHeight}
     on:scroll={onUserScroll}
     class:interrupt
+    class:reverse={reverseScroll}
     class={`scrollable-list ${rootSelector}`}>
-    <slot />
+    <slot {labelObserver} />
 </div>
 
 {#if !readonly}
@@ -526,6 +628,11 @@
     .scrollable-list {
         @include message-list();
         background-color: var(--currentChat-msgs-bg);
+
+        &.reverse {
+            display: flex;
+            flex-direction: column-reverse;
+        }
 
         &.interrupt {
             overflow-y: hidden;